+++ sqldelight-gradle-plugin/src/test/kotlin/app/cash/sqldelight/tests/MigrationTest.kt
@@ -342,4 +342,22 @@ class MigrationTest {
       """.trimMargin(),
     )
   }
+
+  @Test fun `successful migration works properly when initial version is not 1`() {
+    val output = GradleRunner.create()
+      .withCommonConfiguration(File("src/test/migration-success-initial-version-not-one"))
+      .withArguments("clean", "check", "verifyMainDatabaseMigration", "--stacktrace")
+      .build()
+
+    assertThat(output.output).contains("BUILD SUCCESSFUL")
+  }
+
+  @Test fun `compilation succeeds when verifyMigrations is set to true and initial version is not 1`() {
+    val output = GradleRunner.create()
+      .withCommonConfiguration(File("src/test/migration-success-initial-version-not-one"))
+      .withArguments("clean", "generateMainDatabaseInterface", "--stacktrace")
+      .build()
+
+    assertThat(output.output).contains("BUILD SUCCESSFUL")
+  }
 }